fix a warning
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@10672 d7cf8633-e32d-0410-b094-e92efae38249
::stable-branch::merge=1eb56b522ecd9252242470d388c0ad7272f3f26d

+++ vid_wgl.c
@@ -2198,7 +2198,7 @@ int IN_JoystickGetAxisNum(int ControlListType)
 	int i;
 
 	for (i = 0; i < JOY_MAX_AXES; i++)
-		if (dwAxisMap[i] == ControlListType)
+		if (dwAxisMap[i] == (DWORD) ControlListType)
 			return i;
 	return -1;
 }
